backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-camel-case==1.4.2
drf-spectacular==0.27.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
